pet-care-backend/requirements.txt
fastapi
uvicorn
pydantic
requests
openai
python-dotenv
sqlalchemy
pypdf
faiss-gpu
sentence-transformers